Added collision normal detection & .RAW file reading
Added functions for reading a .RAW format file and either outputting
data internally into the program or externally to a .txt file.

Added functions for finding the direction of the colliding face for each
of the following collision types:
* Sphere <-> Sphere
* Sphere <-> AABB (untested)
* AABB <-> AABB

The above functions will draw a line pointing in the direction of the
colliding face from the position of the collider, when in debug mode.

NOTE: test.txt is a test output file made when testing a fileIO
function. A .TXT file must already exist to be used, so I'm including it
in the commit for convenience.

Also, I rebuilt the project, so I'm committing the new object files too
(yay!).