Added Game Object and other minor changes.
Added the GameObject META structure for grouping related things together
(such as colliders to models / etc.)

Note: The player GameObject is linked to the camera, so if you want to
detach them, comment out the gameObjectSetPos() function. If you want to
turn off rendering of the colliders, then comment out the
gameObjectRender() function, or turn off debug mode all together.

Added an offset to colliders so that they can be moved away from the
relative origin of a GameObject if they are attached to one.

Added implementation of a GameObject in main, for use as a reference
when coding GameObject logic.