Added first phase of physics (plus other changes)
Added the RigidBody structure, more monitoring physics.
See RigidBody.h and .c files for documenetation.

Fixed stutter at the beginning of a camera movement AND
made camera movement 8-directional.

Changed the keys in the CameraControlKeys struct to use a 'ControlKey'
structure to check if the key is currently DOWN or not. The code within
the CameraController.c file that checks key presses has been updated.

Cleaned up the main function and removed collider test cases, and moved
non-drawing code out of the display function (moved to FrameUpdate
function).